requirements.txt
py-cord==2.0.0rc1
mcrcon==0.7.0
loguru==0.6.0
Pillow==9.2.0